pyproject.toml
[project]
name = "mockito"
description = "Spying framework"
authors = [
    { name = "herr kaste", email = "herr.kaste@gmail.com" }
]
dependencies = []
readme = "README.rst"
requires-python = ">=3.8"
license = { text = "MIT" }
dynamic = ["version"]

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.version]
source = "vcs"

[tool.hatch.build.hooks.vcs]
version-file = "mockito/_version.py"

[dependency-groups]
dev = [
    "codespell>=2.2.5",
    "flake8>=5.0.4",
    "ipython>=7.34.0",
    "mypy>=1.4.1",
    "numpy>=1.21.6",
    "pyyaml>=6.0.2",
    "pytest>=7.4.4",
    "types-pyyaml>=6.0.12.20241230",
]
docs = [
    "sphinx>=7.4.7; python_version >= '3.12'",
    "sphinx-autobuild>=2021.3.14; python_version >= '3.12'",
    "furo>=2024.8.6; python_version >= '3.12'",
]
